% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_smvr.R
\name{as_smvr}
\alias{as_smvr}
\alias{as_smvr.default}
\title{Convert to \code{smvr} vector}
\usage{
as_smvr(x, ...)

\method{as_smvr}{default}(x, ...)
}
\arguments{
\item{x}{An object to convert to \code{smvr}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class \code{smvr}.
}
\description{
\code{as_smvr()} is a generic function that converts an object to \code{smvr}
vector. The default method uses \code{vctrs::vec_cast()} to convert the object.
}
\examples{
as_smvr(c("1.0.0", "2.0.0-rc.1", "3.0.0+build.1"))
as_smvr(numeric_version(c("1", "2.3")))
as_smvr(NA)
}
