% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_smvr}
\alias{is_smvr}
\title{Check if an object is a \code{smvr} object}
\usage{
is_smvr(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Indicates whether \code{x} is a \link{smvr} object.
}
\description{
Check if an object is a \code{smvr} object
}
\examples{
is_smvr(smvr(1, 2, 3))
}
