% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_release_identifier.R
\name{new_pre_release_identifier}
\alias{new_pre_release_identifier}
\title{pre_release_identifier: Single pre-release identifier for Semantic Versioning}
\usage{
new_pre_release_identifier(x = character())
}
\arguments{
\item{x}{Character or integer value}
}
\value{
An object of class pre_release_identifier
}
\description{
A class representing a single pre-release identifier (string or integer) for Semantic Versioning 2.0.
}
