% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_release_ids.R
\name{new_pre_release_ids}
\alias{new_pre_release_ids}
\title{pre_release_ids: Pre-release identifiers for Semantic Versioning}
\usage{
new_pre_release_ids(
  id1 = new_pre_release_identifier(),
  id2 = new_pre_release_identifier(""),
  id3 = new_pre_release_identifier(""),
  id4 = new_pre_release_identifier(""),
  id5 = new_pre_release_identifier("")
)
}
\arguments{
\item{id1}{First identifier (character/integer/pre_release_identifier, default: NA)}

\item{id2}{Second identifier (character/integer/pre_release_identifier, default: NA)}

\item{id3}{Third identifier (character/integer/pre_release_identifier, default: NA)}

\item{id4}{Fourth identifier (character/integer/pre_release_identifier, default: NA)}

\item{id5}{Fifth identifier (character/integer/pre_release_identifier, default: NA)}
}
\value{
An object of class pre_release_ids
}
\description{
An rcrd class representing up to 5 pre-release identifiers for Semantic Versioning 2.0.
}
