% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_pre_release_ids}
\alias{parse_pre_release_ids}
\title{Parse pre-release identifiers string into pre_release_ids object}
\usage{
parse_pre_release_ids(x)
}
\arguments{
\item{x}{Character vector (e.g. "alpha.1.2" or "")}
}
\value{
pre_release_ids object
}
\description{
Parse pre-release identifiers string into pre_release_ids object
}
