% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_pre_release_ids_impl}
\alias{parse_pre_release_ids_impl}
\title{Internal implementation of parsing function for pre-release ids}
\usage{
parse_pre_release_ids_impl(x, ..., call = caller_env(), check = TRUE)
}
\arguments{
\item{x}{A character vector representing pre-release identifiers.
Each identifier separated by a dot (\code{.}) will be parsed as a
\link{pre_release_identifier}.}

\item{...}{Ignored.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{check}{\code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the function will
check if the pre-release identifiers are valid according to the
regex pattern. If \code{FALSE}, it will skip the validation step.}
}
\value{
A \link{pre_release_ids} vector.
}
\description{
This is an internal function that parses pre-release identifiers
from a character vector. Only the difference to \code{\link[=parse_pre_release_ids]{parse_pre_release_ids()}}
is that it allows to skip the validation step.
}
\keyword{internal}
