% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_semver}
\alias{parse_semver}
\title{parse_semver: Parse semantic version strings into smvr objects}
\usage{
parse_semver(x)
}
\arguments{
\item{x}{Character vector of semantic version strings}
}
\value{
An smvr vector
}
\description{
Parses a character vector of semantic version strings into smvr objects, following the official semver.org regular expression.
}
