% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_smvr}
\alias{parse_smvr}
\title{parse_smvr: Parse semantic version strings into smvr objects}
\usage{
parse_smvr(x)
}
\arguments{
\item{x}{Character vector of semantic version strings}
}
\value{
An smvr vector
}
\description{
Parses a character vector of semantic version strings into smvr objects, following the official semver.org regular expression.
}
