% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_release_ids.R
\name{pre_release_ids}
\alias{pre_release_ids}
\title{pre_release_ids: Pre-release identifiers for Semantic Versioning}
\usage{
pre_release_ids(
  id1 = pre_release_identifier(),
  id2 = pre_release_identifier(""),
  id3 = pre_release_identifier(""),
  id4 = pre_release_identifier(""),
  id5 = pre_release_identifier("")
)
}
\arguments{
\item{id1}{First identifier (character/integer/pre_release_identifier, default: NA)}

\item{id2}{Second identifier (character/integer/pre_release_identifier, default: NA)}

\item{id3}{Third identifier (character/integer/pre_release_identifier, default: NA)}

\item{id4}{Fourth identifier (character/integer/pre_release_identifier, default: NA)}

\item{id5}{Fifth identifier (character/integer/pre_release_identifier, default: NA)}
}
\value{
An object of class pre_release_ids
}
\description{
An rcrd class representing up to 5 pre-release identifiers for Semantic Versioning 2.0.
}
