% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smvr.R
\name{smvr}
\alias{smvr}
\title{smvr: Semantic Versioning 2.0 vector class}
\usage{
smvr(major = integer(), minor = 0L, patch = 0L, pre_release = "", build = "")
}
\arguments{
\item{major}{Major version (integer)}

\item{minor}{Minor version (integer, default: 0)}

\item{patch}{Patch version (integer, default: 0)}

\item{pre_release}{Pre-release identifiers (default: empty string)}

\item{build}{Build string (default: empty string)}
}
\value{
An object of class smvr
}
\description{
A vctrs-based vector class representing Semantic Versioning 2.0.
}
